UIDATIMP-1383: Override purchase order lines limit setting field allows saving fractional numbers (#1348)
