+++ translations/ui-data-import/en.json
@@ -265,6 +265,7 @@
   "validation.chooseDifferentField": "Please choose a different field",
   "validation.selectAnAction": "Action must be selected before the profile can be saved",
   "validation.chooseAtLeastOneValue": "One or more values must be added before the profile can be saved.",
+  "validation.integer": "Please enter a whole number greater than 0 and less than 1000 to continue",
   "validation.enterAsteriskOrNumeric": "Please enter * or numeric value",
   "validation.enterAsteriskOrOtherNumeric": "Please enter * or other numeric value",
   "validation.enterAsteriskOrAlphanumeric": "Please enter * or alphanumeric value",
